Break out suites management classes into their own file
