+++ lib/librato/rack/configuration.rb
@@ -1,3 +1,5 @@
+
+
 module Librato
   class Rack
     # Holds configuration for Librato::Rack middleware to use.
@@ -127,42 +129,8 @@ module Librato
         end
       end
 
-      class Suites
-        attr_reader :fields
-        def initialize(value)
-          @fields = value.to_s.split(/\s*,\s*/).map(&:to_sym)
-        end
-
-        def include?(field)
-          fields.include?(field)
-        end
-      end
-
-      class SuitesAll
-        def include?(value)
-          true
-        end
-      end
-
-      class SuitesExcept < Suites
-        DEFAULT_SUITES = [:rack]
-
-        def initialize(value)
-          super
-          @fields = DEFAULT_SUITES - @fields
-        end
-
-        def include?(field)
-          fields.include?(field)
-        end
-      end
-
-      class SuitesNone
-        def include?(value)
-          false
-        end
-      end
-
     end
   end
 end
+
+require_relative 'configuration/suites'
