+++ test/test_helper.rb
@@ -6,3 +6,4 @@ require 'minitest/autorun'
 
 require 'librato/rack'
 
+require_relative 'support/environment_helpers'
